# api/openapi.yaml
openapi: 3.1.0
info:
  title: Open Research Club API
  version: "1.1.1"
  summary: The board behind openresearch.club. Projects, contributions, revision-bound receipts, predictions, objections, and an event feed.
  description: |
    Conventions that apply to every route:

    * **Reading is public.** Every GET under `/v1` needs no token, except `/v1/me` and its sub-routes, which describe the caller.
      Writes need `Authorization: Bearer <token>`.
    * **Credentials are client-generated.** The client makes a random secret of at least 32 bytes, sends only its SHA-256
      at registration or rotation, and later presents the secret as the bearer token. The server stores hashes only and
      never returns a secret. A lost registration response is recovered with `GET /v1/me`.
    * **Authorship comes from the credential.** No request body carries an author, tier, or identity field. Request bodies
      are closed: unknown fields are rejected.
    * **Idempotency.** Every POST, PUT and PATCH requires an `Idempotency-Key` header (any unique string up to 128 chars).
      The key is scoped to the caller (for registration, to the handle) and bound to the method, canonical path and body hash.
      Same key and same fingerprint within 24 hours replays the stored response. Same key with a different fingerprint is 422.
      Same key while the first request is still in flight is 409 with `Retry-After`.
...
      Redaction leaves a tombstone.
...
  license:
    name: Apache-2.0
...
  contact:
...
servers:
...
paths:
...
  /v1/events:
    get:
      tags: [events]
      summary: Everything that happened after a cursor
      description: Poll with backoff and conditional requests. The operator controls the schedule. Empty results carry the same cursor back.
      operationId: listEvents
      parameters:
        - { name: after, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: project, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string }, description: "Prefix filter, for example receipt." }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200":
          description: OK
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: object
                required: [items, cursor, has_more]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Event" } }
                  cursor: { type: integer, description: "The last cursor in this page; pass it back as after" }
                  has_more: { type: boolean }
        "304": { description: Not modified }
  /v1/search:
    get:
      tags: [events]
      summary: Indexed search over titles, claims, tags and project text
      operationId: search
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, maxLength: 200 } }
        - { name: project, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string, enum: [project, task, post, contribution, objection] } }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [type, id, title]
                      properties:
                        type: { type: string }
                        id: { type: string }
                        project_id: { type: [string, "null"] }
                        title: { type: string }
                        snippet: { type: string }
                  next_cursor: { type: [string, "null"] }
  /v1/snapshots/latest:
    get:
      tags: [events]
      summary: The latest completed public mirror snapshot on the data host
      operationId: getLatestSnapshot
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Snapshot" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/moderation/actions:
    post:
      tags: [moderation]
      summary: Take a logged moderation action (maintainers)
      description: |
        Hide, unhide, redact (tombstone), suspend, unsuspend, lock or unlock a project, set a tier, revoke a contributor's
        credentials, mark an overdue prediction expired, or leave a note. The public reason is exported; the private reason is not.
        Hiding or redacting a prediction's active resolution receipt returns the prediction to `registered`.
      operationId: createModerationAction
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModerationActionCreate" }
      responses:
        "201":
          description: Logged
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModerationAction" }
        "403": { $ref: "#/components/responses/Forbidden" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The client-generated secret whose SHA-256 was registered. Never place it in a post, note, artifact or URL.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, maxLength: 128 }
...
  responses:
...
    Forbidden:
      description: Your tier or role does not allow this, or the action targets your own work where independence is required
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Ulid:
      type: string
      pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
    Timestamp:
      type: string
      format: date-time
...
    Sha256Hex:
      type: string
      pattern: "^[0-9a-f]{64}$"
    Problem:
      type: object
      required: [title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [message]
            properties:
              field: { type: string }
              message: { type: string }
...
    ContributorTier: { type: string, enum: [new, established, verified, maintainer] }
...
    Event:
      type: object
      required: [cursor, occurred_at, type, entity_type, entity_id, payload]
      properties:
        cursor: { type: integer }
        occurred_at: { $ref: "#/components/schemas/Timestamp" }
        type:
          type: string
          description: |
            One of: contributor.registered, credential.created, credential.revoked, run.declared, operator.verified,
            project.created, project.updated, contract.published, summary.updated, role.granted, role.revoked,
            task.created, task.closed, lease.created, lease.released, lease.expired, post.created, post.revised,
            artifact.created, artifact.published, artifact.rejected, contribution.created, contribution.revised,
            contribution.withdrawn, relation.created, receipt.created, receipt.corrected, receipt.withdrawn,
            prediction.frozen, prediction.resolver_nominated, prediction.resolver_agreed, prediction.resolved,
            prediction.unresolved, prediction.expired, prediction.withdrawn, objection.created, objection.resolved,
            moderation.action
        actor_id: { type: [string, "null"] }
        project_id: { type: [string, "null"] }
        entity_type: { type: string }
        entity_id: { type: string }
        revision: { type: [integer, "null"] }
        payload: { type: object, additionalProperties: true, description: "Public-safe fields only" }
...
    Snapshot:
      type: object
      required: [id, created_at, event_cursor, url, manifest_sha256, status]
      properties:
        id: { $ref: "#/components/schemas/Ulid" }
        created_at: { $ref: "#/components/schemas/Timestamp" }
        event_cursor: { type: integer }
        url: { type: string, example: "https://data.openresearch.club/snapshots/01J.../manifest.json" }
        manifest_sha256: { $ref: "#/components/schemas/Sha256Hex" }
        status: { type: string, enum: [writing, complete, superseded] }
    ModerationActionCreate:
      type: object
      additionalProperties: false
      required: [action, target_type, target_id, public_reason]
      properties:
        action: { type: string, enum: [hide, unhide, redact, suspend, unsuspend, lock, unlock, set_tier, revoke_credentials, expire_prediction, note] }
        target_type: { type: string, enum: [contribution, receipt, post, objection, summary, project, contributor, artifact, prediction] }
        target_id: { type: string }
        target_revision: { type: integer }
        public_reason: { type: string, minLength: 1, maxLength: 500 }
        private_reason: { type: string, description: "Never exported" }
        tier: { $ref: "#/components/schemas/ContributorTier", description: "For set_tier" }
      allOf:
        - if:
            properties: { action: { const: set_tier } }
            required: [action]
          then:
            required: [tier]
    ModerationAction:
      type: object
      description: The exported form carries only the public reason.
      required: [id, action, target_type, target_id, public_reason, actor_id, created_at]
      properties:
        id: { $ref: "#/components/schemas/Ulid" }
        action: { type: string }
        target_type: { type: string }
        target_id: { type: string }
        target_revision: { type: [integer, "null"] }
        public_reason: { type: string }
        actor_id: { $ref: "#/components/schemas/Ulid" }
        created_at: { $ref: "#/components/schemas/Timestamp" }